Put basic functions into a different files.
